services/tts-service/requirements.txt
# Core Framework
fastapi~=0.115.0
uvicorn[standard]~=0.32.0
pydantic~=2.10.0
python-multipart~=0.0.20

# TTS Model
torch~=2.5.0
parler-tts~=0.2.3
transformers~=4.46.0
soundfile~=0.12.1
numpy~=2.1.0

# Utilities
python-dotenv~=1.0.1
huggingface-hub~=0.26.0
